+++ generator.c
@@ -4,7 +4,7 @@ void matrix_generator(int nthreads) {
     if (nthreads == 1) {
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
-                matrix[i][j] = (float)(rand() % (RANGE * 2) - RANGE) / SCALE;
+                matrix[i][j] = (double)(rand() % (RANGE * 2) - RANGE) / SCALE;
             }
         }
     }
@@ -12,7 +12,7 @@ void matrix_generator(int nthreads) {
 #pragma omp parallel for num_threads(nthreads)
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
-                matrix[i][j] = (float)(rand() % (RANGE * 2) - RANGE) / SCALE;
+                matrix[i][j] = (double)(rand() % (RANGE * 2) - RANGE) / SCALE;
             }
         }
     }
@@ -21,13 +21,13 @@ void matrix_generator(int nthreads) {
 void vector_generator(int nthreads) {
     if (nthreads == 1) {
         for (int i = 0; i < SIZE; i++) {
-            vec[i][0] = (float)(rand() % (RANGE * 2) - RANGE) / SCALE;
+            vec[i][0] = (double)(rand() % (RANGE * 2) - RANGE) / SCALE;
         }
     }
     else {
 #pragma omp parallel for num_threads(nthreads)
         for (int i = 0; i < SIZE; i++) {
-            vec[i][0] = (float)(rand() % (RANGE * 2) - RANGE) / SCALE;
+            vec[i][0] = (double)(rand() % (RANGE * 2) - RANGE) / SCALE;
         }
     }
 }
